Create Task 1
In this code, I first load the dataset from the provided URL using Pandas. Then we split the data into input (X) and target (y) variables, where X represents the study hours and y represents the scores.

Next, we split the data into training and testing sets using train_test_split() function from scikit-learn. We allocate 80% of the data for training and 20% for testing, using a random state of 42 for reproducibility.

We create a LinearRegression model from scikit-learn, train it using the training data, and then make predictions on the test set using predict() method.

Finally, we evaluate the model's performance using mean squared error (MSE) and coefficient of determination (R^2 score). We also demonstrate how to predict the score for a given study hours value (9.25 in this example).